feat: rm command best practices (#2729)

+++ lgsm/functions/command_dev_detect_glibc.sh
@@ -70,8 +70,8 @@ do
 			echo -e ""
 			echo -e "All required GLIBC versions"
 			cat "${tmpdir}/detect_glibc_${glibc_check_var}.tmp" | sort | uniq | sort -r --version-sort
-			rm "${tmpdir}/detect_glibc_${glibc_check_var}.tmp"
-			rm "${tmpdir}/detect_glibc_files_${glibc_check_var}.tmp"
+			rm -f "${tmpdir:?}/detect_glibc_${glibc_check_var}.tmp"
+			rm -f "${tmpdir:?}/detect_glibc_files_${glibc_check_var}.tmp"
 		else
 			fn_print_information_nl "glibc is not required"
 		fi
@@ -84,7 +84,7 @@ echo -e "Final glibc Requirement"
 echo -e "================================="
 if [ -f "${tmpdir}/detect_glibc_highest.tmp" ]; then
 	cat "${tmpdir}/detect_glibc_highest.tmp" | sort | uniq | sort -r --version-sort | head -1
-	rm "${tmpdir}/detect_glibc_highest.tmp"
+	rm -f "${tmpdir:?}/detect_glibc_highest.tmp"
 else
 	fn_print_information_nl "glibc is not required"
 fi
